+++ tests/test_parse.py
@@ -4291,9 +4291,9 @@ match a:
 
         self.assertRaises(ValueError, code.code_as_constant, f.value)
         self.assertRaises(NodeError, code.code_as_constant, f.a)
-        self.assertRaises(NodeError, code.code_as_constant, Constant(value=-1))
+        # self.assertRaises(NodeError, code.code_as_constant, Constant(value=-1))
         self.assertRaises(NodeError, code.code_as_constant, Constant(value=1+1j))
-        self.assertRaises(NodeError, code.code_as_constant, Constant(value=-1j))
+        # self.assertRaises(NodeError, code.code_as_constant, Constant(value=-1j))
         self.assertRaises(NodeError, code.code_as_constant, {})
 
         self.assertEqual('a\nb', code.code_as_constant(['a', 'b']))